refactor(otgfs): usb common code
This change extracted all the common USB code that will be shared
between otgfs and usbhs
